# server/swagger.yaml
openapi: 3.0.0
paths:
  /xam:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - App
  /xam/videos/storage/clean-junk:
    post:
      operationId: VideosController_cleanStorageJunk
      parameters: []
      responses:
        '201':
          description: ''
      tags: &ref_0
        - Videos
  /xam/videos/queue/clear-all:
    post:
      operationId: VideosController_clearAllQueueAndTasks
      parameters: []
      responses:
        '201':
          description: ''
      tags: *ref_0
  /xam/videos/upload:
    post:
      operationId: VideosController_uploadVideo
      parameters: []
      responses:
        '201':
          description: ''
      tags: *ref_0
  /xam/videos:
    get:
      operationId: VideosController_listProjects
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_0
  /xam/videos/{id}:
    get:
      operationId: VideosController_getProject
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
    delete:
      operationId: VideosController_deleteProject
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /xam/videos/{id}/stream:
    get:
      operationId: VideosController_streamVideo
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /xam/videos/{id}/retry:
    post:
      operationId: VideosController_retryPipeline
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_0
  /xam/videos/{id}/retranscribe:
    post:
      operationId: VideosController_retranscribePipeline
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_0
  /xam/videos/{id}/reset:
    post:
      operationId: VideosController_resetPipeline
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_0
  /xam/videos/{id}/progress:
    get:
      operationId: VideosController_sseProgress
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /xam/subtitles/batch/extract:
    post:
      operationId: SubtitlesController_extractBatch
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchExtractDto'
      responses:
        '201':
          description: ''
      tags: &ref_1
        - Subtitles
  /xam/subtitles/batch/merge:
    post:
      operationId: SubtitlesController_mergeBatch
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeSubtitlesDto'
      responses:
        '201':
          description: ''
      tags: *ref_1
  /xam/subtitles/project/{id}/export:
    get:
      operationId: SubtitlesController_exportSubtitles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: format
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/SubtitleExportFormat'
      responses:
        '200':
          description: ''
      tags: *ref_1
  /xam/subtitles/{id}/cues:
    post:
      operationId: SubtitlesController_createCue
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCueDto'
      responses:
        '201':
          description: ''
      tags: *ref_1
  /xam/subtitles/{id}/cues/{cueId}:
    patch:
      operationId: SubtitlesController_updateCue
      parameters:
        - name: cueId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCueDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
    delete:
      operationId: SubtitlesController_deleteCue
      parameters:
        - name: cueId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /xam/subtitles/{id}/cues/regenerate:
    post:
      operationId: SubtitlesController_regenerateCues
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateCuesDto'
      responses:
        '201':
          description: ''
      tags: *ref_1
  /xam/subtitles/{id}/style:
    patch:
      operationId: SubtitlesController_updateStyle
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStyleDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
  /xam/fonts:
    get:
      operationId: FontsController_getFonts
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Fonts
  /xam/videos/{id}/export:
    post:
      operationId: ExportController_createExport
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportDto'
      responses:
        '201':
          description: ''
      tags: &ref_2
        - Export
  /xam/exports/{jobId}/progress:
    get:
      operationId: ExportController_sseProgress
      parameters:
        - name: jobId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
  /xam/exports/{jobId}/download:
    get:
      operationId: ExportController_downloadExport
      parameters:
        - name: jobId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
info:
  title: XAM-PROD BACKEND
  description: XAM-PROD Application Backend API
  version: 1.0.0
  contact:
    name: SUBHAM MAITY
    url: https://github.com/Subham-Maity
    email: maitysubham4041@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
tags:
  - name: subham-xam
    description: Developer of this app
servers:
  - url: http://localhost:3336
    description: Local server (development)
  - url: http://210.89.44.183:3336/xam
    description: Staging server (testing)
  - url: http://210.89.44.183:3336/xam
    description: Production server (live)
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    SubtitleExportFormat:
      type: string
      enum:
        - srt
        - text
    BatchExtractDto:
      type: object
      properties:
        videoProjectIds:
          description: Array of video project IDs to extract subtitles for
          example:
            - vid_1
            - vid_2
          type: array
          items:
            type: string
        format:
          default: srt
          allOf:
            - $ref: '#/components/schemas/SubtitleExportFormat'
        language:
          type: string
          description: Declared speech language (e.g. hi, hi-roman, en, ur, auto)
          example: hi
      required:
        - videoProjectIds
    MergeSubtitlesDto:
      type: object
      properties:
        videoProjectIds:
          description: Array of video project IDs to merge subtitles from
          example:
            - vid_1
            - vid_2
          type: array
          items:
            type: string
        format:
          default: srt
          allOf:
            - $ref: '#/components/schemas/SubtitleExportFormat'
      required:
        - videoProjectIds
    CreateCueDto:
      type: object
      properties: {}
    UpdateCueDto:
      type: object
      properties: {}
    RegenerateCuesDto:
      type: object
      properties: {}
    UpdateStyleDto:
      type: object
      properties: {}
    CreateExportDto:
      type: object
      properties: {}